+++ api/python/quilt3/admin/util.py
@@ -1,3 +1,5 @@
+import typing as T
+
 from .. import _graphql_client
 from . import exceptions, types
 
@@ -9,7 +11,26 @@ def handle_errors(result: _graphql_client.BaseModel) -> _graphql_client.BaseMode
 
 
 def handle_user_mutation(result: _graphql_client.BaseModel) -> types.User:
-    return types.User(**handle_errors(result).model_dump())
+    return types.User.from_gql(handle_errors(result))
+
+
+def raise_invalid_input(error: _graphql_client.InvalidInputSelection) -> T.NoReturn:
+    first_error = error.errors[0] if error.errors else None
+    if first_error is None:
+        raise exceptions.InvalidInputError(error)
+    if first_error.name == "PolicyDoesNotExist":
+        raise exceptions.PolicyNotFoundError()
+    if first_error.name == "PolicyTitleConflict":
+        raise exceptions.PolicyTitleExistsError(error)
+    if first_error.name == "PolicyArnConflict":
+        raise exceptions.PolicyArnExistsError(error)
+    if first_error.name == "RoleHasTooManyPoliciesToAttach":
+        raise exceptions.RoleTooManyPoliciesError(error)
+    raise exceptions.InvalidInputError(error)
+
+
+def raise_operation_error(error: _graphql_client.OperationErrorSelection) -> T.NoReturn:
+    raise exceptions.OperationError(error)
 
 
 def get_client():